+++ backend/src/neuroagent/app/main.py
@@ -33,6 +33,7 @@ from neuroagent.app.dependencies import (
 )
 from neuroagent.app.middleware import strip_path_prefix
 from neuroagent.app.routers import qa, rate_limit, storage, threads, tools
+from neuroagent.executor import WasmExecutor
 from neuroagent.mcp import MCPClient
 
 LOGGING = {
@@ -129,12 +130,23 @@ async def lifespan(fastapi_app: FastAPI) -> AsyncContextManager[None]:  # type: 
         fastapi_app.state.accounting_session_factory = session_factory
 
         async with MCPClient(config=app_settings.mcp) as mcp_client:
-            # trigger dynamic tool generation - only done once - it is cached
-            _ = fastapi_app.dependency_overrides.get(
-                get_mcp_tool_list, get_mcp_tool_list
-            )(mcp_client, app_settings)
-            fastapi_app.state.mcp_client = mcp_client
-            yield
+            with WasmExecutor(
+                additional_imports=[
+                    "numpy",
+                    "pandas",
+                    "file:./node_modules/pyodide/plotly-6.3.1-py3-none-any.whl",
+                    "pydantic",
+                    "scikit-learn",
+                    "scipy",
+                ]
+            ) as sandbox:
+                fastapi_app.state.python_sandbox = sandbox
+                # trigger dynamic tool generation - only done once - it is cached
+                _ = fastapi_app.dependency_overrides.get(
+                    get_mcp_tool_list, get_mcp_tool_list
+                )(mcp_client, app_settings)
+                fastapi_app.state.mcp_client = mcp_client
+                yield
 
     # Cleanup connections
     if engine: